chore: rename h2 proxy
